+++ Package.swift
@@ -19,11 +19,16 @@ let package = Package(
     dependencies: [
         .package(url: "https://github.com/simibac/ConfettiSwiftUI", .upToNextMajor(from: "1.0.0")),
         .package(url: "https://github.com/danielsaidi/SwiftUIKit.git", .upToNextMajor(from: "3.4.0")),
+        .package(url: "https://github.com/danielsaidi/SystemNotification", .upToNextMajor(from: "0.7.0")),
     ],
     targets: [
         .target(
             name: "KankodaKit",
-            dependencies: ["ConfettiSwiftUI", "SwiftUIKit"]
+            dependencies: [
+                "ConfettiSwiftUI",
+                "SwiftUIKit",
+                "SystemNotification"
+            ]
         ),
         .testTarget(
             name: "KankodaKitTests",